[doc2] Fix issue where a copy of a shared document in user home becomes unreachable
This takes care of an issue where a copy of a shared document ends up in the original owners home directory and becomes unreachable by the user making the copy. With this change these copied files are now put in the home directory of the user that makes the copy instead. If the file to be copied exists in a shared folder the copy will end up in that same shared folder, a user being able to copy such a document implies that the user is also able to browse that directory.