DWX-19423: Remove permission settings from Docker image (#128)
We need to:
- run with a UID > 1000000
- disallow changing the UID on the run
- set root filesystem as read-only

Plus today Hue has today sudo privileges, that has to go immediately:
```
 echo "${HUEUSER} ALL=(root) NOPASSWD:ALL" > /etc/sudoers.d/${HUEUSER} 
```

A security context in kubernetes can set these restrictions, today what we have on most of our pods look something like this:
```
      securityContext:
        runAsUser: 1000
        runAsGroup: 1000
        fsGroup: 1000
```

With following security standards, it should look like this:

```
      securityContext:
        runAsUser: 1000001
        runAsGroup: 1000001
        fsGroup: 1000001
        allowPrivilegeEscalation: false
        readOnlyRootFilesystem: true
```

The one we have today works because we also set the UID to 1000 in our Dockerfiles. This is why doing any chown to 1000 can work. If we were to change the pod’s UID to 1000001, it would not be able to access files belonging to 1000. One solution for that would be to change the UID to 1000001 in all our Dockerfiles, but then we need to keep Dockerfiles and these security contexts in sync, and keep chowning in the future.
The better way to handle this is via `fsGroup`, which sets the files we add to the dockerfile to belong to the specified group. Unless we directly chown the file, the running user would have access to those files if we set `runAsUser` and `runAsGroup`. 

So with all these set, we can omit chowns in the Dockerfile and rely on the pod’s user to be picked up.

mkdir is also similar to chown in this regard as the owner of the dir would be the user we create it with. The reason you have to chown today a lot in your Dockerfiles is because you’re creating these users as root (default user from UBI minimal). 

The reason I’m initially testing on OpenShift is because it can provide this out of the box. OpenShift sets `runAsUser`, `runAsGroup` , and fsGroup automatically to a random UI over 1000000. It does not provide `readOnlyRootFilesystem` but could help discover where else we need write permissions. This is why I changed a few paths to `/var/run/hue`, writing temporary data or files necessary to run the pod will need a common location with write access, while we can keep the config directories as read-only. These files are created on startup and match the pod’s lifecycle, so changing them is safe in terms of upgrades. In Kubernetes we would mount an emptyDir to `/var/run/hue` and similar locations, giving the path with write access to the pod.

The reason the [DWX PR](https://github.infra.cloudera.com/CDH/dwx/pull/10213) is talking about removing the need for an OpenShift SCC, aka Security Context Constraint is because in the initial days of PvC releases we bypassed this default security mechanism with a custom SCC. An SCC constraints the security contexts, forces us to only include in them what is specified. The default SCC (restricted-v2 ) beside [many other restrictions](https://docs.openshift.com/container-platform/4.11/authentication/managing-security-context-constraints.html), forces the pods to run with a UID > 1000000, we would not be able to set the security context I added above.

Change-Id: Ib4a174c5c9cf57011ccc22806471f36c8cbfeef9