Moved django-babel into hue repo
django-babel has been out of maintenance for a while. We forked it
several years ago and have been updating it to support newer and newer
django versions. This commit brings the whole django-babel code into
our repo and makes it simpler to maintain it in future. This will also
allow us to make django upgrade changes in lock-step with the hue app.