[security] SAML login issues with csp enabled
pysaml by default gives a template which has unsafe inline code. django saml bypasses that if we use a custom template. i have defined that custom template and added a request object.

Uses the post binding form to add nonce, if the nonce is available in the request. if csp_nonce is disabled in the setting it won't be added here.

Co-authored-by: Mohammed Tabraiz <tabraiz@cloudera.com>