[django] fixed was_modified_since by removing size param (#4087)
In django4.1 the size param has been removed. Justification from django
looks good.

Removed outdated handling of length parameter to If-Modified-Since header
The length parameter is not described in RFC-7232 and it's against
HTTP/1.0 and HTTP/1.1 specifications. It was an old and unofficial
extension set by some ancient versions of IE.