[django] Upgraded djangosaml2 to 1.9.3
We earlier had djangosaml2-0.18 in ext-py3 with three commits on top of
it.
1. [Fix SAML TemplateDoesNotExist](fa1b914)
2. [SAML login issues with csp enabled](864283c)
3. [Fix for removal of django.conf.urls.url in Django4](e22fab5)

#2 and #3 are not needed as the fixes are already there in the newer
version.
I did not run into any issues yet regarding `TemplateDoesNotExist`. So,
I am assuming that the fix might no longer be needed on our upgraded
django4+djangosaml2v1.9.3 combination. But, if we see the
`TemplateDoesNotExist` error ever, then we can consider bringing
djangosaml2 again in our repo and applying this fix.