[beeswax] Do not error on invalid session in request
If the frontend sends a request with session that is no longer in the
db or malformed, the current `Session.object.get()` throws an error.
The change to using `filter().first()` would get the session if found
in db or `None` rather than throwing an exception. This would let the
query proceed with a new session getting created.