[cookies] cookies to expire at browser closing
Our `sessionid` cookie has a 14 days of expiry. This means that if a
user forgets to click logout button and instead just closes the browser
window, they remain signed in. This is a security hole, especially if
the computer is used by multiple people. Someone else can gain access
to the previous user's data.

This commits sets the default age of `sessionid` and `csrftoken` to `0`.
This will make these cookies non-persistent and held only in memory.
These cookies will get cleared on browser restart.

In addition, this commit also enforces the same for google analytics
cookies.